Revert "search: refactor SearchResult into CommitSearchResult and RepoSearchResult (#34804)"
This reverts commit a55d11a18186c377cbd95cbab4763d12bbb85bb7.
