--- client/search-ui/src/components/CommitSearchResult.tsx
@@ -1,99 +1,0 @@
-import React from 'react'
-
-import { displayRepoName } from '@sourcegraph/shared/src/components/RepoFileLink'
-import { RepoIcon } from '@sourcegraph/shared/src/components/RepoIcon'
-import { ResultContainer } from '@sourcegraph/shared/src/components/ResultContainer'
-import { SearchResultStar } from '@sourcegraph/shared/src/components/SearchResultStar'
-import { PlatformContextProps } from '@sourcegraph/shared/src/platform/context'
-import { CommitMatch, getCommitMatchUrl, getRepositoryUrl } from '@sourcegraph/shared/src/search/stream'
-import { formatRepositoryStarCount } from '@sourcegraph/shared/src/util/stars'
-// eslint-disable-next-line no-restricted-imports
-import { Timestamp } from '@sourcegraph/web/src/components/time/Timestamp'
-import { Link, useIsTruncated } from '@sourcegraph/wildcard'
-
-import { CommitSearchResultMatch } from './CommitSearchResultMatch'
-
-import styles from './SearchResult.module.scss'
-
-interface Props extends PlatformContextProps<'requestGraphQL'> {
-    result: CommitMatch
-    repoName: string
-    icon: React.ComponentType<{ className?: string }>
-    onSelect: () => void
-    openInNewTab?: boolean
-    containerClassName?: string
-}
-
-export const CommitSearchResult: React.FunctionComponent<Props> = ({
-    result,
-    icon,
-    repoName,
-    platformContext,
-    onSelect,
-    openInNewTab,
-    containerClassName,
-}) => {
-    /**
-     * Use the custom hook useIsTruncated to check if overflow: ellipsis is activated for the element
-     * We want to do it on mouse enter as browser window size might change after the element has been
-     * loaded initially
-     */
-    const [titleReference, truncated, checkTruncation] = useIsTruncated()
-
-    const renderTitle = (): JSX.Element => {
-        const formattedRepositoryStarCount = formatRepositoryStarCount(result.repoStars)
-        return (
-            <div className={styles.title}>
-                <RepoIcon repoName={repoName} className="text-muted flex-shrink-0" />
-                <span
-                    onMouseEnter={checkTruncation}
-                    className="test-search-result-label ml-1 flex-shrink-past-contents text-truncate"
-                    ref={titleReference}
-                    data-tooltip={(truncated && `${result.authorName}: ${result.message.split('\n', 1)[0]}`) || null}
-                >
-                    <>
-                        <Link to={getRepositoryUrl(result.repository)}>{displayRepoName(result.repository)}</Link>
-                        {' › '}
-                        <Link to={getCommitMatchUrl(result)}>{result.authorName}</Link>
-                        {': '}
-                        <Link to={getCommitMatchUrl(result)}>{result.message.split('\n', 1)[0]}</Link>
-                    </>
-                </span>
-                <span className={styles.spacer} />
-                <Link to={getCommitMatchUrl(result)}>
-                    <code className={styles.commitOid}>{result.oid.slice(0, 7)}</code>{' '}
-                    <Timestamp date={result.authorDate} noAbout={true} strict={true} />
-                </Link>
-                {formattedRepositoryStarCount && (
-                    <>
-                        <div className={styles.divider} />
-                        <SearchResultStar />
-                        {formattedRepositoryStarCount}
-                    </>
-                )}
-            </div>
-        )
-    }
-
-    const renderBody = (): JSX.Element => (
-        <CommitSearchResultMatch
-            key={result.url}
-            item={result}
-            platformContext={platformContext}
-            openInNewTab={openInNewTab}
-        />
-    )
-
-    return (
-        <ResultContainer
-            icon={icon}
-            collapsible={false}
-            defaultExpanded={true}
-            title={renderTitle()}
-            resultType={result.type}
-            onResultClicked={onSelect}
-            expandedChildren={renderBody()}
-            className={containerClassName}
-        />
-    )
-}
