+++ client/search-ui/src/components/index.ts
@@ -1,3 +1,2 @@
 export * from './SyntaxHighlightedSearchQuery'
-export * from './CommitSearchResult'
-export * from './RepoSearchResult'
+export * from './SearchResult'
